+++ AppleDocumentationPackage/Sources/App/Pages/TechnologyDetailPage/Views/Renderer/ContentViews/ParagraphView.swift
@@ -1,0 +1,17 @@
+import SwiftUI
+import AppleDocumentation
+import UIComponent
+
+struct ParagraphView: View {
+    let paragraph: DocumentData.ParagraphItem
+    let headingLevel: Int?
+
+    var body: some View {
+        Text { next in
+            for text in paragraph.texts {
+                next(text)
+            }
+        }
+        .headingLevel(headingLevel)
+    }
+}
